Refactor CSV import plugin
- Separate Sky Dolly CSV parser
