+++ src/Plugin/src/Import/KMLImport/src/FlightRadar24KMLParser.cpp
@@ -95,10 +95,10 @@ FlightRadar24KMLParser::~FlightRadar24KMLParser() noexcept
 #endif
 }
 
-// FlightRadar24 KML files (are expected to) have 3 Placemarks, with:
-// - <Point> Takeoff airpart
-// - <Point> Destination airport
-// - <gx:Track> timestamps (<when>) and positions (<gx:coord>)s
+// FlightRadar24 KML files (are expected to) have one "Route" folder with <Placemarks> containing
+// - <description> - HTML snippet containing speed and heading
+// - <Timestamp> timestamps
+// - <Point> - the coordinates of the track
 void FlightRadar24KMLParser::parse(QDateTime &firstDateTimeUtc, QDateTime &lastDateTimeUtc, QString &flightNumber) noexcept
 {
     d->trackData.clear();
@@ -192,8 +192,6 @@ bool FlightRadar24KMLParser::parseDescription() noexcept
     QRegularExpressionMatch match = d->speedRegExp.match(description);
     if (match.hasMatch()) {
         FlightRadar24KMLParserPrivate::TrackItem trackItem;;
-
-
         trackItem.speed = match.captured(1).toInt();
         const int pos = match.capturedEnd();
         match = d->headingRegExp.match(description, pos);
